+++ meteor-app/.storybook/meteor/session.js
@@ -1,0 +1,7 @@
+const s = {}
+
+export const Session = {
+  get: name => s[name],
+  set: (name, val) => (s[name] = val),
+  clear: name => delete s[name]
+}