+++ meteor-app/imports/ui/utils/nav.js
@@ -1,4 +1,17 @@
-import { Session } from 'meteor/session'
+//
+// This module acts as a replacement for meteor/session
+const sessionStore = {}
+
+const context = {
+  get: name => sessionStore[name],
+  set: (name, val) => (sessionStore[name] = val),
+  clear: name => delete sessionStore[name],
+  goHome: () => {
+    const mode = context.get('mode')
+    const dest = destinations[mode] || destinations.default
+    return dest
+  }
+}
 
 const destinations = {
   kiosk: '/kiosk',
@@ -7,8 +20,4 @@ const destinations = {
   default: '/'
 }
 
-export default goHome = () => {
-  const mode = Session.get('mode')
-  const dest = destinations[mode] || destinations.default
-  return dest
-}
+export default context